requirements.txt
django
gunicorn
django-heroku
django-crispy-forms
django-guardian
dj-database-url
